+++ chat/message-rating/src/MessageRating/MessageRating.styles.ts
@@ -16,17 +16,6 @@ export const buttonContainerStyles = css`
   gap: ${spacing[100]}px;
 `;
 
-const baseActiveStyles = css`
-  &::before {
-    background-color: initial;
-  }
-`;
-
-const getActiveStyles = (isActive: boolean) =>
-  cx({
-    [baseActiveStyles]: isActive,
-  });
-
 const baseHiddenStyles = css`
   display: none;
 `;
@@ -35,11 +24,3 @@ export const getHiddenStyles = (isHidden: boolean) =>
   cx({
     [baseHiddenStyles]: isHidden,
   });
-
-export const getIconButtonStyles = ({
-  isActive,
-  isHidden,
-}: {
-  isActive: boolean;
-  isHidden: boolean;
-}) => cx(getActiveStyles(isActive), getHiddenStyles(isHidden));