+++ chat/message/src/components/ToolCard/ToolCard.stories.tsx
@@ -11,7 +11,9 @@ import LeafyGreenProvider from '@leafygreen-ui/leafygreen-provider';
 
 import { State, ToolCard, ToolCardProps } from './index';
 
-const chips = [{ label: 'MongoDB', glyph: <DatabaseIcon /> }];
+const chips = [
+  { label: 'docdb-elastic.amazonaws.com:27017', glyph: <DatabaseIcon /> },
+];
 const titles = {
   [State.Canceled]: 'Canceled list-databases',
   [State.Error]: 'Failed running list-databases',
@@ -78,10 +80,21 @@ const renderToolCardChildren = () => (
     <ToolCard.ExpandableContent>
       {`#### ARGUMENTS
 
-\`\`\`javascript
+\`\`\`json
 {
-  foo: 'bar',
-  baz: 'qux',
+  "database": "sample_mflix",
+  "collection": "embedded_movies",
+  "pipeline": [
+    {
+      "$search": {
+        "index": "movie_search_index",
+        "text": {
+          "query": "science fiction",
+          "path": "plot"
+        }
+      }
+    }
+  ]
 }
 \`\`\``}
     </ToolCard.ExpandableContent>