+++ packages/combobox/src/Combobox.stories.tsx
@@ -5,7 +5,7 @@ import {
   type StoryMetaType,
   StoryType,
 } from '@lg-tools/storybook-utils';
-import { StoryFn } from '@storybook/react';
+import { StoryContext, StoryFn } from '@storybook/react';
 import { userEvent, within } from '@storybook/test';
 
 import { Button } from '@leafygreen-ui/button';
@@ -150,15 +150,15 @@ const meta: StoryMetaType<typeof Combobox> = {
 
 export default meta;
 
-export const LiveExample: StoryFn<ComboboxProps<boolean>> = (
-  args: ComboboxProps<boolean>,
-) => {
+export const LiveExample: StoryFn<ComboboxProps<boolean>> = args => {
   return (
     <>
       {/* Since Combobox doesn't fully refresh when `multiselect` changes, we need to explicitly render a different instance */}
       {args.multiselect ? (
+        // @ts-ignore - multiselect check ensures props match ComboboxProps<true>
         <Combobox key="multi" {...args} multiselect={true} />
       ) : (
+        // @ts-ignore - multiselect check ensures props match ComboboxProps<false>
         <Combobox key="single" {...args} multiselect={false} />
       )}
     </>
@@ -265,6 +265,7 @@ export const MultiSelectNoIcons: StoryFn<ComboboxProps<boolean>> = (
   args: ComboboxProps<boolean>,
 ) => {
   return (
+    // @ts-expect-error - args will have multiselect=true from storybook controls
     <Combobox {...args} multiselect={true}>
       {getComboboxOptions(false)}
     </Combobox>
@@ -299,20 +300,20 @@ export const InitialLongComboboxOpen = {
       </Combobox>
     );
   },
-  play: async ctx => {
+  play: async (ctx: StoryContext) => {
     const { findByRole } = within(ctx.canvasElement.parentElement!);
     const trigger = await findByRole('combobox');
     userEvent.click(trigger);
   },
   decorators: [
-    (StoryFn, _ctx) => (
+    (Story: StoryFn, _ctx: StoryContext) => (
       <div
         className={css`
           height: 100vh;
           padding: 0;
         `}
       >
-        <StoryFn />
+        <Story />
       </div>
     ),
   ],