+++ packages/combobox/src/ComboboxOption/ComboboxOption.tsx
@@ -1,20 +1,14 @@
 import React, { useCallback, useContext, useMemo } from 'react';
 
-import { cx } from '@leafygreen-ui/emotion';
 import { useForwardedRef, useIdAllocator } from '@leafygreen-ui/hooks';
 import { InputOption, InputOptionContent } from '@leafygreen-ui/input-option';
 import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
+import { getNodeTextContent } from '@leafygreen-ui/lib';
 
 import { ComboboxContext } from '../ComboboxContext';
-import { ComboboxSize } from '../types';
 import { wrapJSX } from '../utils';
 
-import {
-  displayNameStyle,
-  largeStyles,
-  multiselectIconLargePosition,
-  multiselectIconPosition,
-} from './ComboboxOption.styles';
+import { getInputOptionStyles } from './ComboboxOption.styles';
 import {
   ComboboxOptionProps,
   InternalComboboxOptionProps,
@@ -43,6 +37,7 @@ export const InternalComboboxOption = React.forwardRef<
       value,
       onClick,
       disabled = false,
+      'aria-label': ariaLabel,
       ...rest
     }: InternalComboboxOptionProps,
     forwardedRef,
@@ -106,17 +101,14 @@ export const InternalComboboxOption = React.forwardRef<
         ref={optionRef}
         highlighted={isFocused}
         disabled={disabled}
-        aria-label={displayName}
+        aria-label={ariaLabel || getNodeTextContent(displayName) || value}
         darkMode={darkMode}
-        className={cx(
-          {
-            [largeStyles]: size === ComboboxSize.Large,
-            [multiselectIconPosition]: multiSelectWithoutIcons,
-            [multiselectIconLargePosition]:
-              multiSelectWithoutIcons && size === ComboboxSize.Large,
-          },
+        className={getInputOptionStyles({
+          size,
+          isSelected,
+          isMultiselectWithoutIcons: multiSelectWithoutIcons,
           className,
-        )}
+        })}
         onClick={handleOptionClick}
         onKeyDown={handleOptionClick}
       >
@@ -125,9 +117,13 @@ export const InternalComboboxOption = React.forwardRef<
           rightGlyph={rightGlyph}
           description={description}
         >
-          <span id={optionTextId} className={displayNameStyle(isSelected)}>
-            {wrapJSX(displayName, inputValue, 'strong')}
-          </span>
+          {typeof displayName === 'string' ? (
+            <span id={optionTextId}>
+              {wrapJSX(displayName, inputValue, 'strong')}
+            </span>
+          ) : (
+            displayName
+          )}
         </InputOptionContent>
       </InputOption>
     );