+++ packages/pagination/src/Pagination/Summary/Summary.types.ts
@@ -1,6 +1,9 @@
+import { LgIdProps } from '@leafygreen-ui/lib';
+
 import { PaginationProps } from '../Pagination.types';
 
 export type SummaryProps<T extends number = number> = Pick<
   PaginationProps<T>,
   'itemsPerPage' | 'currentPage' | 'numTotalItems' | 'className'
->;
+> &
+  LgIdProps;